fix(rig): keep the non-persistent rig path audit non-fatal in strict mode
The boot audit's doc comment promised it "only ever warns — refusing to
load a city because one rig path looks doomed would brick a running city
that a warning could have saved". It did the opposite.

The warning rides Provenance.Warnings into `gc start`, where
splitStrictConfigWarnings treats every warning it does not recognize as
fatal and strict mode is on by default. So the guard bricked exactly the
cities it was meant to warn:

    gc init --default-provider claude --no-start ./mycity   # exit 0
    gc rig add /dev/shm/rig --allow-ephemeral               # exit 0
    gc start --controller --dry-run                         # exit 1
      gc start: strict: rig "rig" is bound to /dev/shm/... does not
      survive a restart

`gc rig add` created a city that `gc start` then refused to boot, through
the PR's own opt-out flag. A legacy /tmp-rooted city hit the same wall on
upgrade, with --no-strict hidden and refused outside the legacy standalone
path, and a running controller's config reload hit it too.

Exempt the audit in IsNonFatalSiteBindingWarning, keyed on a fragment that
is also interpolated into the message so the exemption cannot drift from
the text it exempts. Both halves of the contract are now pinned:
TestNonPersistentRigPathWarningIsExemptFromStrictMode derives the real
message from the producer and asserts the classifier accepts it, and
TestStartBootsCityWithAllowedEphemeralRig runs rig add -> config load ->
strict split end to end over real text.

Also in this pass:

- probe_linux.go read the superblock magic into a signed, arch-dependent
  type. On 386/arm Statfs_t.Type is int32, so ramfs (0x858458f6)
  sign-extended to a negative int64 and matched nothing. Key the table on
  uint32 and truncate through it on every arch.
- rig_add_path_durability_test.go proved its own precondition with
  pathdurability.Classify, so gutting Classify skipped all three e2e tests
  and the package reported ok. Establish the precondition from the kernel
  instead and fail loudly once it holds.
- The same tests used a fixed /dev/shm name that collided across
  concurrent `go test` runs; take a unique scratch dir per run.
- `gc rig add --allow-ephemeral` against a remote city silently ignored
  the flag while the server's refusal told the user to pass it. Refuse it
  loudly, symmetric with the --git-url and --request-id guards.
- isRealFilesystem duplicated isOSFileSystem in the same package; reuse it.

+++ internal/pathdurability/durability.go
@@ -66,7 +66,7 @@ const (
 // overlayfs covers the container rootfs, which is what makes an in-pod path
 // like /var/tmp ephemeral even though it is not /tmp — the reason a prefix
 // denylist cannot do this job.
-var ephemeralFilesystems = map[int64]string{
+var ephemeralFilesystems = map[uint32]string{
 	magicTmpfs:     "tmpfs",
 	magicRamfs:     "ramfs",
 	magicOverlayfs: "overlayfs",
